pyCodeGen p7: added datatype xsd:list, a simpleType that contains a list of other simple or basetypes. Changes in SimpleTypesMapping and ConvertTo../From.. User-defined typdefs added to gsl-file (via pyCodeGen.py) to provide interfaces to the API.

+++ include/OSCLib/Data/OSCP/MDIB/Measurement.h
@@ -59,8 +59,8 @@ public:
 	Measurement & setMeasurementUnit(const CodedValue & value);
 	CodedValue getMeasurementUnit() const;
 
-	Measurement & setMeasuredValue(const decimal & value);
-	decimal getMeasuredValue() const;
+	Measurement & setMeasuredValue(const double & value);
+	double getMeasuredValue() const;
 
 private:
 	std::shared_ptr<CDM::Measurement> data;
